refactor: move the simulation to the lib
the lib folder contains logical components, the compiler, simulation
and so on. the app folder is for the react/angular parts only. this
will make it easier to extract the raw logic at a later point if
needed
